abstract `<NavbarDropdownItem />` from `NavbarDropdownMenu.tsx` to a new file
